requirements.txt
fastapi
uvicorn
pandas
matplotlib
scipy
openai
python-multipart
python-dotenv
requests
duckdb